+++ examples/jiti.js
@@ -1,21 +1,13 @@
 const path = require('path')
-const { transformSync } = require('@babel/core')
 
 module.exports = require('jiti')(__filename, {
   cache: false,
   debug: false,
-  transform (opts) {
-    const _opts = {
-      babelrc: false,
-      configFile: false,
-      compact: false,
-      retainLines: typeof opts.retainLines === 'boolean' ? opts.retainLines : true,
-      filename: '',
-      cwd: '/',
+  transformOptions: {
+    ts: true,
+    retainLines: true,
+    babel: {
       plugins: [
-        [require('@babel/plugin-transform-modules-commonjs'), { allowTopLevelThis: true }],
-        [require('@babel/plugin-transform-typescript')],
-        [require('babel-plugin-dynamic-import-node'), { noInterop: true }],
         [require('babel-plugin-global-define'), {
           __DEV__: true,
           __BROWSER__: false
@@ -24,22 +16,11 @@ module.exports = require('jiti')(__filename, {
           root: '.',
           extensions: ['.ts'],
           alias: {
-            '^vue-meta$': path.resolve(__dirname, '../src/')
+            '^vue-meta$': path.resolve(__dirname, '../src/'),
+            '^vue-meta/ssr$': path.resolve(__dirname, '../src/ssr')
           }
         }]
       ]
-    }
-
-    try {
-      return transformSync(opts.source, _opts).code || ''
-    } catch (err) {
-      return 'exports.__JITI_ERROR__ = ' + JSON.stringify({
-        filename: opts.filename,
-        line: (err.loc && err.loc.line) || 0,
-        column: (err.loc && err.loc.column) || 0,
-        code: err.code && err.code.replace('BABEL_', '').replace('PARSE_ERROR', 'ParseError'),
-        message: err.message.replace('/: ', '').replace(/\(.+\)\s*$/, '')
-      })
     }
   }
 })