+++ examples/vue-router/browser.js
@@ -1,7 +1,4 @@
-import { createApp } from 'vue'
-import { App, createRouter, metaManager } from './main'
+import { createApp } from './main'
 
-const app = createApp(App)
-app.use(createRouter('/vue-router'))
-app.use(metaManager)
-app.mount('#app')
+const { app, router } = createApp('/vue-router')
+router.isReady().then(() => app.mount('#app'))