+++ src/ssr.ts
@@ -1,16 +1,12 @@
 import type { App } from 'vue'
 import type { SSRContext } from '@vue/server-renderer'
 
-export async function renderToStringWithMeta (app: App): Promise<[string, SSRContext]> {
-  const { renderToString } = await import('@vue/server-renderer')
-
-  const ctx: SSRContext = {}
-
-  const html = await renderToString(app, ctx)
-
+export async function renderMetaToString (app: App, ctx: SSRContext = {}): Promise<SSRContext> {
   // TODO: better way of determining whether meta was rendered with the component or not
   if (!ctx.teleports || !ctx.teleports.head) {
-    const teleports = app.config.globalProperties.$metaManager.render()
+    const { renderToString } = await import('@vue/server-renderer')
+
+    const teleports = app.config.globalProperties.$metaManager?.render()
     await Promise.all(teleports.map((teleport: any) => renderToString(teleport, ctx)))
   }
 
@@ -24,5 +20,5 @@ export async function renderToStringWithMeta (app: App): Promise<[string, SSRCon
     }
   }
 
-  return [html, ctx]
+  return ctx
 }